Filled in missing fields
After discussions with @andrew-coleman, we agreed that adding value and position
consistently can only help improve error messages.  The issue was that it would break
existing tests and appear to be a regression.  So we've updated the tests to reflect
the new output.  It is important to emphasize that these changes did not impact any
semantic elements of the AST, they only added information that was previously
missing.