I changed to target to es2015 to help generate better coverage
If we target es5, the generators get rewritten and those rewritten
generators lose the istanbul related comments for some cases.  As
a result, we get what looks like bad coverage.

Switching to es2015 fixed the coverage and did not break any of the
testing or debugging.  So I'm going to stick with that as the baseline
development target.  We can always include an additional target to
generate es5 as well.