All previously passing tests running
This includes fixes for #163 (including detection of the new semantic error).
Note that I've changed the code the evaluates object construction to map
more closely to how v1.5+ did it.

Still having issues with the flattening cases, but I wanted to checkpoint here
since all the previously passing tests are working.