Cleaned up various corner cases
These cases had to do with distinguishing between expressions
that evaluated to `undefined` vs. expressions that evaluated to
`[]`.  Part of this was related to how we iterated over empty
boxes (and adding some special handling for that) and some of
it was just about contextual information that indicated whether
we were definitely dealing with arrays.

The big change here was that boxing an empty array now
returns a `ubox`.  This had some downstream implications
in the test sweet which have not been resolved.

All passing cases pass again now and we still have 1003
cases passing overall and 25 failing (ignoring test cases that
we skipped because of outstanding issues in the issue tracker).