Narrowed handling of arrays returned by functions
Previously I had changed the semantics so that if a function
returned an array, it was represented as an ArrayBox.  But I
changed it so that if it returns an *empty* array it is represented
as an ArrayBox (vs. a VoidBox) but if it has values, it just calls
`boxValue`.